feat: 랜덤 번호 생성 함수 분리
